stage 2

+++ resources/js/helpers.ts
@@ -1,9 +1,9 @@
 import { CalendarDate, CalendarDateTime, ZonedDateTime, getLocalTimeZone } from '@internationalized/date';
 
 
-export const formatDate = (date: CalendarDate, opts :Intl.DateTimeFormatOptions = {weekday: 'long', day: 'numeric', month: 'long', year: 'numeric' }) =>
+export const formatDate = (date: CalendarDate, opts :Intl.DateTimeFormatOptions = {weekday: 'long', day: 'numeric', month: 'short', year: 'numeric' }) =>
         new Intl.DateTimeFormat('pt-PT', opts).format(date.toDate(getLocalTimeZone()));
 
 export  const formatDateTime = (date: CalendarDateTime |ZonedDateTime, opts :Intl.DateTimeFormatOptions =
-{ weekday: 'long',day: 'numeric', month: 'long', year: 'numeric', hour: '2-digit', minute: '2-digit' }) =>
+{ weekday: 'long',day: 'numeric', month: 'short', year: 'numeric', hour: '2-digit', minute: '2-digit' }) =>
     new Intl.DateTimeFormat('pt-PT', opts).format(date.toDate(getLocalTimeZone()));
